requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
torch==2.2.2
torchvision==0.17.2
elasticsearch==7.17.3
httpx==0.27.0
Pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.12
